Changed the width and height of sprites to be float rather than int

+++ include/sprite.h
@@ -16,9 +16,9 @@ typedef struct {
 } Sprite;
 
 Sprite sprite_init(jmp_buf error, vec3 pos, unsigned int scale, char *texture_location,
-		vec2 translation[], vec2 spr_num[], int instance_amount, int width, int height);
+		vec2 translation[], vec2 spr_num[], int instance_amount, float width, float height);
 void sprite_draw(Sprite *sprite, unsigned int program, int instance_amount);
 void sprite_delete(Sprite *sprite);
 
 
-void sprite_generate_model(Model *model, int width, int height);
+void sprite_generate_model(Model *model, float width, float height);
